requirements.txt
markdown
google-generativeai
Flask
Flask-Cors
